requirements.txt
dune-client==1.3.0
pandas==2.2.0
python-dotenv==1.0.1 